Module shared: Add a new interface GameNavigator and rename the FrameNavigator interface to WindowNavigator. Both are implemented in the module view.
Module model:
Call the interface of SessionData to add and get scores and let SessionData handle the ScoreManager.
Call the interface of ScoreManager directly to write scores to a file.

Module controller:
Delegate writing scores to a file to the GameNavigator so that special view worker threads can be used.
Add a window listener to the start window and have it call GameNavigator's exitGame() on close.